+++ Enigma/Enigma.Core/Services/CoreServices.cs
@@ -9,6 +9,7 @@ using Enigma.Core.Communication;
 using Enigma.Core.Configuration;
 using Enigma.Core.Handlers;
 using Enigma.Core.Persistency;
+using Enigma.Core.Research;
 using Enigma.Core.Research.Services;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
@@ -60,6 +61,7 @@ public static class CoreServices
         serviceCollection.AddTransient<IDateTimeCalc, DateTimeCalc>();
         serviceCollection.AddTransient<IDateTimeHandler, DateTimeHandler>();
         serviceCollection.AddTransient<IDateTimeValidator, DateTimeValidator>();
+        serviceCollection.AddTransient<IDeclinationParallelsCounting, DeclinationParallelsCounting>();
         serviceCollection.AddTransient<IDeclMidpointsHandler, DeclMidpointsHandler>();
         serviceCollection.AddTransient<IDefaultConfiguration, DefaultConfiguration>();
         serviceCollection.AddTransient<IDefaultProgConfiguration, DefaultProgConfiguration>();
@@ -88,7 +90,9 @@ public static class CoreServices
         serviceCollection.AddTransient<IObliqueLongitudeHandler, ObliqueLongitudeHandler>();
         serviceCollection.AddTransient<IObliquityCalc, ObliquityCalc>();
         serviceCollection.AddTransient<IObliquityHandler, ObliquityHandler>();
+        serviceCollection.AddTransient<IOccupiedMidpointsDeclinationCounting, OccupiedMidpointsDeclinationCounting>();
         serviceCollection.AddTransient<IOccupiedMidpointsFinder, OccupiedMidpointsFinder>();
+        serviceCollection.AddTransient<IOobCounting, OobCounting>();
         serviceCollection.AddTransient<IParallelsHandler, ParallelsHandler>();
         serviceCollection.AddTransient<IPdDataFromToRdbmsHandler, PdDataFromToRdbmsHandler>();
         serviceCollection.AddTransient<IPdDataFromToPersistableConverter, PdDataFromToPersistableConverter>();